+++ src/Entity/Track.php
@@ -8,6 +8,7 @@ use App\Entity\Track\OptimizedPoint;
 use App\Entity\Track\Version;
 use App\Entity\User\User;
 use App\Entity\Video\Youtube;
+use App\EntityTraits\DescriptionTrait;
 use App\EntityTraits\NameTrait;
 use App\EntityTraits\SendByTrait;
 use DateTime;
@@ -21,6 +22,7 @@ class Track implements CreatedAtInterface
 {
     use NameTrait;
     use SendByTrait;
+    use DescriptionTrait;
 
     public const TYPE_CYCLING = 1;
     public const TYPE_HIKING = 2;
@@ -91,16 +93,6 @@ class Track implements CreatedAtInterface
      * @ORM\Column(type="datetime")
      */
     private $createdAt;
-
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     */
-    private $descriptionEn;
-
-    /**
-     * @ORM\Column(type="string", nullable=true)
-     */
-    private $descriptionBg;
 
     /**
      * @ORM\Column(type="integer")
@@ -398,43 +390,6 @@ class Track implements CreatedAtInterface
     public function getCreatedAt(): DateTime
     {
         return $this->createdAt;
-    }
-
-    public function getDescriptionEn(): ?string
-    {
-        return $this->descriptionEn;
-    }
-
-    public function setDescriptionEn(string $text = null): void
-    {
-        $this->descriptionEn = $text;
-    }
-
-    public function getDescriptionBg(): ?string
-    {
-        return $this->descriptionBg;
-    }
-
-    public function setDescriptionBg(string $text = null): void
-    {
-        $this->descriptionBg = $text;
-    }
-
-    /**
-     * Get localized name.
-     * If localized name doesn't exists - fallback to EN
-     */
-    public function getDescription(string $locale): ?string
-    {
-        if ($locale === 'bg' && $this->descriptionBg !== null) {
-            return $this->descriptionBg;
-        }
-
-        if ($this->descriptionEn === null) {
-            return $this->descriptionBg;
-        }
-
-        return $this->descriptionEn;
     }
 
     /**
