+++ src/Form/Type/Place.php
@@ -6,6 +6,7 @@ use App\Translations;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\NumberType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Validator\Constraints\Range;
@@ -64,6 +65,32 @@ class Place extends AbstractType
                 'scale' => 8,
                 'label' => 'Longitude',
                 'constraints' => $coordinatesConstraints,
+            ]
+        );
+
+        $builder->add(
+            'descriptionBg',
+            TextareaType::class,
+            [
+                'required' => false,
+                'label' => 'Description in Bulgarian',
+                'attr' => [
+                    'data-html' => 'wysiwyg',
+                    'style' => 'min-height: 300px',
+                ]
+            ]
+        );
+
+        $builder->add(
+            'descriptionEn',
+            TextareaType::class,
+            [
+                'required' => false,
+                'label' => 'Description in English',
+                'attr' => [
+                    'data-html' => 'wysiwyg',
+                    'style' => 'min-height: 300px',
+                ]
             ]
         );
     }
